requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
yfinance>=0.2.28
newsapi-python>=0.2.7
transformers>=4.30.0
torch>=2.0.0
joblib>=1.3.0
psycopg2-binary>=2.9.0
requests>=2.31.0
textblob>=0.18.0
feedparser>=6.0.0
finnhub-python>=2.4.0
lxml>=4.9.0
vaderSentiment>=3.3.2
quantstats>=0.0.62
TA-Lib>=0.4.28
pdfplumber>=0.11.4
beautifulsoup4>=4.12.3
APScheduler>=3.10.4
langdetect>=1.0.9
deep-translator>=1.11.4
python-dotenv>=1.0.1
newspaper3k>=0.2.8

